fix(mouth/voa): move flag helper out of layout.tsx — Next App Router closed export list
Vercel build on ac5afac39 failed with TS2344: `export function
isGarudaVoaPublicEnabled()` inside layout.tsx violates the closed export
set Next's App Router permits from a layout file (default/metadata/dynamic/
generateStaticParams/...). The route-type shim that rejects the extra
export only exists at real `next build` time (.next/types/app/**/layout.ts)
— `tsc --noEmit` genuinely cannot see this class of error, which is why the
prior round measured clean and still broke on Vercel.

Fix: extracted the helper to its own module, src/app/visa/voa/flag.ts, with
no export-list restriction. layout.tsx now exports exactly three things —
dynamic, metadata, default — and imports isGarudaVoaPublicEnabled from
./flag. The flag is still read at CALL time inside the function (never
hoisted to a module-level const, which would re-bake the value into a
static render); dynamic = "force-dynamic" stays on the layout itself, since
that's route-segment config Next only honors at the segment level.

The 12 test cases carry over split by concern: 9 flag-parsing cases moved
to flag.test.ts (unit tests of pure string logic), 3 rendering/gate cases
stay in layout.test.tsx (notFound() called / not called against the mocked
next/navigation).

Verified: `npx tsc --noEmit -p .` clean, `npx vitest run
src/app/visa/voa/flag.test.ts src/app/visa/voa/layout.test.tsx` 12/12
green, and — per the actual failure mode above — a full `npm run build`
run to completion: exit code 0 (captured directly, not through a piped
tail), zero `error TS`/`Failed to compile` lines in the log, and
/visa/voa + /visa/voa/[hash] both listed as ƒ Dynamic in the route output.
